Split DisplayableVisualizationTypeData from config factory
