+++ src/DetailsView/components/target-page-changed-view.tsx
@@ -6,7 +6,7 @@ import * as commonStaticStyles from 'DetailsView/components/static-content-commo
 import * as styles from 'DetailsView/components/target-page-changed-view.scss';
 import * as React from 'react';
 
-import { DisplayableVisualizationTypeData } from '../../common/configs/visualization-configuration-factory';
+import { DisplayableVisualizationTypeData } from '../../common/types/displayable-visualization-type-data';
 import { NamedFC } from '../../common/react/named-fc';
 import { VisualizationType } from '../../common/types/visualization-type';
