+++ src/common/types/displayable-visualization-type-data.ts
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+export interface DisplayableVisualizationTypeData {
+    title: string;
+    subtitle?: JSX.Element;
+    enableMessage: string;
+    toggleLabel: string;
+    linkToDetailsViewText: string;
+}
